requirements.txt
pyserial==3.5
serial-tool==0.0.1
matplotlib==3.9.2
python-dotenv==1.0.1